Moved files into folders in componeents and worked on starting to import css into App.css in style folder
