+++ project-3/client/src/App.js
@@ -5,9 +5,9 @@ import Home from "./components/Home/index";
 import SignUp from "./components/SignUp/signup"
 import SignIn from "./components/SignIn/signin"
 // import CurrentExpenses from "./components/currentExpenses";
-import ExpenseForm from "./components/expenseForm";
+import ExpenseForm from "./components/Expense-Form/expenseForm";
 import Profile from "./components/Profile/profile";
-import "./app.css";
+import "./style/App.css"
 import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
 import API from "./utils/api";
