--- project-3/client/src/app.css
@@ -1,8 +1,0 @@
-* {
-    outline: transparent!important;
-    text-decoration: none!important;
-}
-
-body {
-    padding-top: 77px;
-}
