--- project-3/client/src/assets/style.css
@@ -1,3 +1,0 @@
-#root {
-    padding-top: 77px;
-}
