+++ project-3/client/src/components/Expense-Form/expenseForm.js
@@ -1,9 +1,8 @@
 import * as React from "react";
 // import { Router } from 'react-router'
 // stateless functional component that returns html (jsx)
-import "../assets/additionalcss/css/signup.css";
-import "../assets/additionalcss/css/mbr-additional.css";
-import "./expense-form.css";
+import "../../assets/additionalcss/css/signup.css";
+import "../../assets/additionalcss/css/mbr-additional.css";
 class AddExpense extends React.Component{
   constructor(props){
     super(props);
