--- project-3/client/src/components/expense-form.css
@@ -1,43 +1,0 @@
-/* button {
-    border-radius: 5px;
-    padding: 8px 25px 8px 25px;
-    color: pink;
-}
-
-#burger-icon {
-max-height: 25px!important;
-max-width: 25px!important;
-} */
-.textform {
-    display: inline !important;
-    font-display: black;
-    place-content: black;
-    text-align: center !important;
-    height: 45px !important;
-    width: 600px;
-    color: black !important;
-}
-select{
-    text-align: center !important;
-}
-button {
-    border-radius: 5px;
-    padding: 8px 25px 8px 25px;
-    color: pink;
-}
-input[type="text"], input[type="email"], input[type="password"] {
-    height: 36px !important;
-    width: 600px;
-}
-.expenseform {
-    height: 64px !important;
-    width: 600px !important;
-}
-.signin{
-    height: 36px !important;
-    width: 400px !important;
-}
-/* #burger-icon {
-max-height: 25px!important;
-max-width: 25px!important;
-} */
